requirements.txt
gunicorn==21.2.0
Bootstrap-Flask==2.4.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
flask-cors==5.0.0
attrs~=23.2.0
Jinja2~=3.1.3
packaging~=23.2
future~=1.0.0
SQLAlchemy~=2.0.25
bcrypt~=4.2.1
pyjwt~=2.4.0
python-dotenv~=1.0.1
countryinfo~=0.1.2
pycountry~=24.6.1
timezonefinder~=6.5.9
geopy~=2.4.1
CurrencyConverter~=0.18.7
